requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10.1
django-extensions==1.7.4
djangorestframework==3.4.7
easy-thumbnails==2.3
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
oauthlib==2.0.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
requests==2.11.1
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
twython==3.4.0
wcwidth==0.1.7
